requirements.txt
Django>=4.2
django-allauth>=0.52.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
requests>=2.32